requirements.txt
grpcio 
grpcio-tools
asgiref
Django
django-grpc-framework